pgbench: Remove stray use of "float" math.
Oversight in commit 4a87f308b33457670f9ab4bc622678e5d285b9c2.

Fabien COELHO

+++ contrib/pgbench/pgbench.c
@@ -3077,7 +3077,7 @@ threadRun(void *arg)
 				/* generate and show report */
 				int64 count = 0;
 				int64 run = now - last_report;
-				float tps, total_run, latency;
+				double tps, total_run, latency;
 
 				for (i = 0 ; i < nstate ; i++)
 					count += state[i].cnt;
@@ -3107,7 +3107,7 @@ threadRun(void *arg)
 				/* generate and show report */
 				int64 count = 0;
 				int64 run = now - last_report;
-				float tps, total_run, latency;
+				double tps, total_run, latency;
 
 				for (i = 0 ; i < progress_nclients ; i++)
 					count += state[i].cnt;
